# Improve BuilderNode pooling significantly and add better comments / documentation