--- a/builder_test.go
+++ b/builder_test.go
@@ -66,6 +66,7 @@ func TestBuilderSimple(t *testing.T) {
 	if err != nil {
 		t.Errorf("got error closing set builder: %v", err)
 	}
+	b.Reset(ioutil.Discard)
 }
 
 func TestBuilderSharedPrefix(t *testing.T) {
@@ -238,12 +239,12 @@ func BenchmarkBuilder(b *testing.B) {
 
 	b.ResetTimer()
 
+	builder, err := New(ioutil.Discard, nil)
+	if err != nil {
+		b.Fatalf("error creating builder: %v", err)
+	}
 	for i := 0; i < b.N; i++ {
-
-		builder, err := New(ioutil.Discard, nil)
-		if err != nil {
-			b.Fatalf("error creating builder: %v", err)
-		}
+		builder.Reset(ioutil.Discard)
 		err = insertStrings(builder, dataset, randomThousandVals)
 		if err != nil {
 			b.Fatalf("error inserting thousand words: %v", err)