--- a/registry.go
+++ b/registry.go
@@ -37,16 +37,25 @@ func newRegistry(p *builderNodePool, tableSize, mruSize int) *registry {
 	return rv
 }
 
+var notEmptyCount = 0
+
 func (r *registry) Reset() {
 	var empty registryCell
 	for i := range r.table {
+		if r.table[i].node != nil {
+			notEmptyCount++
+		}
 		r.builderNodePool.Put(r.table[i].node)
 		r.table[i] = empty
 	}
 }
 
+var entryCount = 0
+
 func (r *registry) entry(node *builderNode) (bool, int, *registryCell) {
+	entryCount++
 	if len(r.table) == 0 {
+		panic("bad")
 		return false, 0, nil
 	}
 	bucket := r.hash(node)
@@ -77,6 +86,9 @@ func (r *registry) hash(b *builderNode) int {
 
 type registryCache []registryCell
 
+var matchCount = 0
+var insertCount = 0
+
 func (r registryCache) entry(node *builderNode, pool *builderNodePool) (bool, int, *registryCell) {
 	if len(r) == 1 {
 		if r[0].node != nil && r[0].node.equiv(node) {
@@ -90,9 +102,11 @@ func (r registryCache) entry(node *builderNode, pool *builderNodePool) (bool, in
 		if r[i].node != nil && r[i].node.equiv(node) {
 			addr := r[i].addr
 			r.promote(i)
+			matchCount++
 			return true, addr, nil
 		}
 	}
+
 	// no match
 	last := len(r) - 1
 	pool.Put(r[last].node)