The existing implementation was missing a lot of places where BuilderNodes should be returned to the pool, which was causing the pool to leak massively and allocations would be very high when building FSTs. This addresses that issue, improves the comments / documentation, and also sets more reasonable default values for the pooling.

before
```
goos: darwin
goarch: amd64
pkg: github.com/m3db/vellum
BenchmarkBuilder-8           2000        649700 ns/op      269256 B/op        5832 allocs/op
PASS
ok      github.com/m3db/vellum    1.381s
Success: Benchmarks passed.
```

after
```
goos: darwin
goarch: amd64
pkg: github.com/m3db/vellum
BenchmarkBuilder-8   	    3000	    465007 ns/op	     737 B/op	       5 allocs/op
PASS
ok  	github.com/m3db/vellum	1.465s
Success: Benchmarks passed.
```