@sreekanth-cb 2019-03-28
Hi @richardartoul,

Apologies for the delayed response, and really appreciate the upstream contribution efforts!

This is very cool! It's gonna take a bit to absorb what you did here, but the #'s are really exciting!

One other thing is, we actually do code reviews over in another system, called Gerrit (same as what android, chrome, golang, libreoffice and others use).

It's running here... http://review.couchbase.org/ and here's the vellum drill down: http://review.couchbase.org/#/q/project:vellum

One remaining item is that we ask that contributors sign the Couchbase CLA, see https://github.com/couchbaselabs/vellum/blob/master/CONTRIBUTING.md

We shall get this reviewed, verified and get is merged.