@mschoch 2021-02-12
Closing this now, as it is covered by #28